smdk4210-common: Improve Settings search UX
Currently, the preference titles are very vague and confusing,
with no way for the user to understand what the preferences
do at a glance (e.g. in Settings search).

Make the purpose of the preferences more explicit through the titles.

Also, the current drawable is colored white, when it should be
colored green. This results in a near-invisible icon in the
Settings search results.

Color the drawable the appropriate teal instead.

Change-Id: I2dac9a73e8689f14d676b41922e0eba364c6500f